Fix: Search Criteria Component
